+++ pygnssutils/gnssdump.py
@@ -12,8 +12,10 @@ Created on 26 May 2022
 # pylint: disable=line-too-long eval-used
 
 import sys
+import os
 from socket import socket
 from queue import Queue
+from datetime import datetime
 from io import TextIOWrapper, BufferedWriter
 from serial import Serial
 import pynmeagps.exceptions as nme
@@ -24,6 +26,7 @@ from pygnssutils.exceptions import ParameterError
 from pygnssutils.globals import (
     VALCKSUM,
     GET,
+    ALL_PROTOCOL,
     UBX_PROTOCOL,
     NMEA_PROTOCOL,
     RTCM3_PROTOCOL,
@@ -34,10 +37,12 @@ from pygnssutils.globals import (
     FORMAT_HEX,
     FORMAT_HEXTABLE,
     FORMAT_PARSEDSTRING,
+    FORMAT_JSON,
     VERBOSITY_LOW,
     VERBOSITY_MEDIUM,
+    LOGLIMIT,
 )
-from pygnssutils.helpers import hextable, protocol
+from pygnssutils.helpers import hextable, protocol, format_json
 from pygnssutils.helpstrings import GNSSDUMP_HELP
 
 
@@ -61,7 +66,7 @@ class GNSSStreamer:
 
     def __init__(self, **kwargs):
         """
-        Constructor.
+        Context manager constructor.
 
         Example of usage with external protocol handler:
 
@@ -76,16 +81,19 @@ class GNSSStreamer:
         :param int validate: (kwarg) 1 = validate checksums, 0 = do not validate (1)
         :param int msgmode: (kwarg) 0 = GET, 1 = SET, 2 = POLL (0)
         :param int parsebitfield: (kwarg) 1 = parse UBX 'X' attributes as bitfields, 0 = leave as bytes (1)
-        :param int format: (kwarg) output format 1 = parsed, 2 = raw, 4 = hex, 8 = tabulated hex, 16 = parsed as string (1) (can be OR'd)
+        :param int format: (kwarg) output format 1 = parsed, 2 = raw, 4 = hex, 8 = tabulated hex, 16 = parsed as string, 32 = JSON (1) (can be OR'd)
         :param int quitonerror: (kwarg) 0 = ignore errors,  1 = log errors and continue, 2 = (re)raise errors (1)
         :param int protfilter: (kwarg) 1 = NMEA, 2 = UBX, 4 = RTCM3 (7 - ALL)
         :param str msgfilter: (kwarg) comma-separated string of message identities e.g. 'NAV-PVT,GNGSA' (None)
         :param int limit: (kwarg) maximum number of messages to read (0 = unlimited)
         :param int verbosity: (kwarg) log message verbosity 0 = low, 1 = medium, 3 = high (1)
-        :param object errorhandler: (kwarg) either writeable output medium or evaluable expression (None)
+        :param int logtofile: (kwarg) 0 = log to stdout, 1 = log to file '/logpath/gnssdump-timestamp.log' (0)
+        :param int logpath: {kwarg} fully qualified path to logfile folder (".")
+        :param object allhandler: (kwarg) either writeable output medium or evaluable expression (None)
         :param object nmeahandler: (kwarg) either writeable output medium or evaluable expression (None)
         :param object ubxhandler: (kwarg) either writeable output medium or evaluable expression (None)
         :param object rtcmhandler: (kwarg) either writeable output medium or evaluable expression (None)
+        :param object errorhandler: (kwarg) either writeable output medium or evaluable expression (None)
         :raises: ParameterError
         """
         # pylint: disable=raise-missing-from
@@ -94,6 +102,14 @@ class GNSSStreamer:
         self._datastream = kwargs.get("datastream", None)
         self._port = kwargs.get("port", None)
         self._socket = kwargs.get("socket", None)
+        if self._socket is not None:
+            sock = self._socket.split(":")
+            if len(sock) != 2:
+                raise ParameterError(
+                    "socket keyword must be in the format host:port.\nType gnssdump -h for help."
+                )
+            self._socket_host = sock[0]
+            self._socket_port = int(sock[1])
         self._filename = kwargs.get("filename", None)
         if (
             self._datastream is None
@@ -119,48 +135,120 @@ class GNSSStreamer:
             )
             self._msgfilter = kwargs.get("msgfilter", None)
             self._verbosity = int(kwargs.get("verbosity", VERBOSITY_MEDIUM))
+            self._logtofile = int(kwargs.get("logtofile", 0))
+            self._logpath = kwargs.get("logpath", ".")
             self._limit = int(kwargs.get("limit", 0))
             self._parsing = False
             self._stream = None
             self._msgcount = 0
             self._errcount = 0
+            self._validargs = True
+            self._loglines = 0
+            self._stopevent = False
+            self._allhandler = None
+            self._errorhandler = None
+            self._nmeahandler = None
+            self._ubxhandler = None
+            self._rtcmhandler = None
 
-            # protocol handlers can either be writeable output media
-            # (Serial, File, socket or Queue) or an evaluable expression
-            # acceptable output media types:
-            htypes = (Serial, TextIOWrapper, BufferedWriter, Queue, socket)
+            # following is to keep tabs on where we are
+            # in any JSON array for each protocol handler
+            self._jsontop = {
+                ALL_PROTOCOL: True,
+                NMEA_PROTOCOL: True,
+                UBX_PROTOCOL: True,
+                RTCM3_PROTOCOL: True,
+            }
 
-            erh = kwargs.get("errorhandler", None)
-            if isinstance(erh, htypes) or erh is None:
+            self._setup_protocol_handlers(**kwargs)
+
+        except (ParameterError, ValueError, TypeError) as err:
+            self._do_log(
+                f"Invalid input arguments {kwargs}\n{err}\n{GNSSDUMP_HELP}",
+                VERBOSITY_LOW,
+            )
+            self._validargs = False
+
+    def _setup_protocol_handlers(self, **kwargs):
+        """
+        Set up protocol handlers.
+
+        protocol handlers can either be writeable output media
+        (Serial, File, socket or Queue) or an evaluable expression.
+
+        'allhandler' applies to all protocols and overrides
+        individual protocol handlers.
+        """
+
+        htypes = (Serial, TextIOWrapper, BufferedWriter, Queue, socket)
+
+        if "errorhandler" in kwargs:
+            erh = kwargs["errorhandler"]
+            if isinstance(erh, htypes):
                 self._errorhandler = erh
             else:
                 self._errorhandler = eval(erh)
-            nmh = kwargs.get("nmeahandler", None)
-            if isinstance(nmh, htypes) or nmh is None:
+
+        if "allhandler" in kwargs:
+            allh = kwargs["allhandler"]
+            if isinstance(allh, htypes):
+                self._allhandler = allh
+            else:
+                self._allhandler = eval(allh)
+            return
+
+        if "nmeahandler" in kwargs:
+            nmh = kwargs["nmeahandler"]
+            if isinstance(nmh, htypes):
                 self._nmeahandler = nmh
             else:
                 self._nmeahandler = eval(nmh)
-            ubh = kwargs.get("ubxhandler", None)
-            if isinstance(ubh, htypes) or ubh is None:
+
+        if "ubxhandler" in kwargs:
+            ubh = kwargs["ubxhandler"]
+            if isinstance(ubh, htypes):
                 self._ubxhandler = ubh
             else:
                 self._ubxhandler = eval(ubh)
-            rth = kwargs.get("rtcmhandler", None)
-            if isinstance(rth, htypes) or rth is None:
+
+        if "rtcmhandler" in kwargs:
+            rth = kwargs["rtcmhandler"]
+            if isinstance(rth, htypes):
                 self._rtcmhandler = rth
             else:
                 self._rtcmhandler = eval(rth)
 
-        except ValueError:
-            raise ParameterError(f"Invalid parameter(s).\n{GNSSDUMP_HELP}")
+    def __enter__(self):
+        """
+        Context manager enter routine.
+        """
 
-    def run(self, **kwargs):
+        return self
+
+    def __exit__(self, exc_type, exc_value, exc_traceback):
+        """
+        Context manager exit routine.
+        """
+
+        self.stop()
+
+    def run(self, **kwargs) -> int:
         """
         Read from provided data stream (serial, file or other stream type).
         The data stream must support a read(n) -> bytes method.
 
         :param int limit: (kwarg) maximum number of messages to read (0 = unlimited)
+        :return: rc 0 = fail, 1 = ok
+        :rtype: int
+        :raises: ParameterError if socket is not in form host:port
         """
+
+        if not self._validargs:
+            return 0
+
+        # if outputting json, add opening tag
+        if self._format == FORMAT_JSON:
+            self._cap_json(1)
 
         self._limit = int(kwargs.get("limit", self._limit))
 
@@ -175,18 +263,28 @@ class GNSSStreamer:
                 self._start_reader()
         elif self._socket is not None:  # socket
             with socket() as self._stream:
-                sock = self._socket.split(":")
-                if len(sock) != 2:
-                    raise ParameterError(
-                        "socket keyword must be in the format host:port.\nType gnssdump -h for help."
-                    )
-                host = sock[0]
-                port = int(sock[1])
-                self._stream.connect((host, port))
+                self._stream.connect((self._socket_host, self._socket_port))
                 self._start_reader()
         elif self._filename is not None:  # binary file
             with open(self._filename, "rb") as self._stream:
                 self._start_reader()
+
+        return 1
+
+    def stop(self):
+        """
+        Shutdown streamer.
+        """
+
+        # if outputting json, add closing tag
+        if self._format == FORMAT_JSON:
+            self._cap_json(0)
+
+        self._stopevent = True
+        mss = "" if self._msgcount == 1 else "s"
+        ers = "" if self._errcount == 1 else "s"
+        msg = f"Streaming terminated, {self._msgcount:,} message{mss} processed with {self._errcount:,} error{ers}.\n"
+        self._do_log(msg, VERBOSITY_MEDIUM)
 
     def _start_reader(self):
         """Create GNSSReader instance."""
@@ -199,7 +297,7 @@ class GNSSStreamer:
             msgmode=self._msgmode,
             parsebitfield=self._parsebitfield,
         )
-        self._do_log(f"\nParsing GNSS data stream from: {self._stream}...\n")
+        self._do_log(f"Parsing GNSS data stream from: {self._stream}...\n")
         self._do_parse()
 
     def _do_parse(self):
@@ -214,7 +312,9 @@ class GNSSStreamer:
 
         try:
 
-            while True:  # loop until EOF, stream timeout or user hits Ctrl-C
+            while (
+                not self._stopevent
+            ):  # loop until EOF, stream timeout or user hits Ctrl-C
 
                 try:
                     (raw_data, parsed_data) = self._reader.read()
@@ -251,6 +351,8 @@ class GNSSStreamer:
                 elif msgprot == RTCM3_PROTOCOL:
                     msgidentity = parsed_data.identity
                     handler = self._rtcmhandler
+                if self._allhandler is not None:
+                    handler = self._allhandler
                 # does it pass the protocol filter?
                 if self._protfilter & msgprot:
                     # does it pass the message identity filter if there is one?
@@ -258,24 +360,24 @@ class GNSSStreamer:
                         if msgidentity not in self._msgfilter:
                             continue
                     # if it passes, send to designated output
-                    self._do_output(raw_data, parsed_data, handler)
+                    self._do_output(msgprot, raw_data, parsed_data, handler)
 
                 if self._limit and self._msgcount >= self._limit:
                     raise EOFError
 
-        except KeyboardInterrupt:  # user hit Ctrl-C
-            self._do_log("user")
         except EOFError:  # end of stream
-            self._do_log("eof")
+            self._do_log("End of file or limit reached", VERBOSITY_LOW)
+            self.stop()
         except Exception as err:  # pylint: disable=broad-except
             self._quitonerror = ERR_RAISE  # don't ignore irrecoverable errors
             self._do_error(err)
 
-    def _do_output(self, raw: bytes, parsed: object, handler: object):
+    def _do_output(self, msgprot: int, raw: bytes, parsed: object, handler: object):
         """
         Output message to terminal in specified format(s) OR pass
         to external protocol handler if one is specified.
 
+        :param int msgprot: protocol 0 = ALL, 1 = NMEA, 2 = UBX, 4 = RTCM
         :param bytes raw: raw (binary) message
         :param object parsed: parsed message
         :param object handler: Queue, socket or protocol handler (NMEA, UBX or RTCM3)
@@ -295,6 +397,8 @@ class GNSSStreamer:
                 print(hextable(raw))
             if self._format & FORMAT_PARSEDSTRING:
                 print(str(parsed))
+            if self._format & FORMAT_JSON:
+                print(format_json(parsed))
             return
 
         # writeable output media (can output one format)
@@ -306,6 +410,16 @@ class GNSSStreamer:
             output = str(raw.hex())
         elif self._format == FORMAT_HEXTABLE:
             output = str(hextable(raw))
+        elif self._format == FORMAT_JSON:
+            # if outputting JSON for this protocol, ensure array
+            # of messages is correctly formatted without trailing
+            # comma at end [{msg1},{msg2},...,[lastmsg]]
+            msgprot = msgprot if self._allhandler is None else 0
+            if self._jsontop[msgprot]:
+                output = format_json(parsed)
+                self._jsontop[msgprot] = False
+            else:
+                output = "," + format_json(parsed)
         else:
             output = raw
         if isinstance(handler, (Serial, TextIOWrapper, BufferedWriter)):
@@ -346,25 +460,56 @@ class GNSSStreamer:
             self._errorhandler(err)
         self._errcount += 1
 
-    def _do_log(self, msg: str, loglevel: int = VERBOSITY_MEDIUM):
+    def _do_log(
+        self,
+        message: str,
+        loglevel: int = VERBOSITY_MEDIUM,
+    ):
         """
-        Log output according to verbosity setting.
+        Write timestamped log message according to verbosity and logfile settings.
 
-        :param str msg: log message
-        :param int loglevel: min verbosity level for this message
+        :param str message: message to log
+        :param int loglevel: log level for this message (0,1,2)
         """
 
-        if msg in ("eof", "user"):
-            mss = "" if self._msgcount == 1 else "s"
-            ers = "" if self._errcount == 1 else "s"
-            pre = (
-                "Streaming terminated by user"
-                if msg == "user"
-                else "End of stream reached"
-            )
-            msg = f"\n\n{pre}, {self._msgcount:,} message{mss} processed with {self._errcount:,} error{ers}.\n"
-        if loglevel <= self._verbosity and self._verbosity > VERBOSITY_LOW:
-            print(msg)
+        msg = f"{datetime.now()}: {message}"
+        if self._verbosity >= loglevel:
+            if self._logtofile:
+                self._cycle_log()
+                with open(self._logpath, "a", encoding="UTF-8") as log:
+                    log.write(msg + "\n")
+                    self._loglines += 1
+            else:
+                print(msg)
+
+    def _cycle_log(self):
+        """
+        Generate new timestamped logfile path.
+        """
+
+        if not self._loglines % LOGLIMIT:
+            tim = datetime.now().strftime("%Y%m%d%H%M%S")
+            self._logpath = os.path.join(self._logpath, f"gnssdump-{tim}.log")
+            self._loglines = 0
+
+    def _cap_json(self, start: int):
+        """
+        Caps JSON file for each protocol handler.
+
+        :param int start: 1 = start, 0 = end
+        """
+
+        for handler in (
+            self._allhandler,
+            self._nmeahandler,
+            self._ubxhandler,
+            self._rtcmhandler,
+        ):
+            if handler is not None:
+                if start:
+                    handler.write('{"GNSS_Messages": [')
+                else:
+                    handler.write("]}")
 
     @property
     def datastream(self) -> object:
@@ -393,10 +538,12 @@ def main():
             sys.exit()
 
     try:
-        gns = GNSSStreamer(**dict(arg.split("=") for arg in sys.argv[1:]))
-        gns.run()
-    except ValueError:
-        raise ParameterError(f"Invalid parameter(s).\n{GNSSDUMP_HELP}")
+
+        with GNSSStreamer(**dict(arg.split("=") for arg in sys.argv[1:])) as gns:
+            gns.run()
+
+    except KeyboardInterrupt:
+        pass
 
 
 if __name__ == "__main__":